feat(test): add spreadsheet of synthetic data - rename test.json to schema.json
