+++ ui/test/README.md
@@ -1,0 +1,5 @@
+# Test Data
+
+This folder contains synthetic data that is used for end-to-end tests of the application.
+
+Synthetic data are generated with [mimesis](https://mimesis.name/en/master/) and are based on the [json schema](../src/schema/test.json) of this project's data.
